+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.4] - 2020-02-11
+### Fixed
+- CDN Provider Status / TTL Bug
+
+## [0.11.3] - 2020-01-30
+### Fixed
+- Batch Domain Error Message
+- Batch Domain Add Operation Log for job
+
 ## [0.11.2] - 2020-01-21
 ### Fixed
 - Batch Domain Error Message